+++ contracts/locked-astroport-vault/src/migrations.rs
@@ -1,5 +1,6 @@
 use crate::state::{self, FeeConfig, CONFIG};
-use cosmwasm_std::{Decimal, DepsMut, StdResult};
+use cosmwasm_std::{Addr, Decimal, DepsMut, StdResult};
+use cw_dex_astroport::AstroportStaking;
 use cw_storage_plus::Item;
 use locked_astroport_vault_0_2_0::state::Config as Config0_2_0;
 
@@ -12,7 +13,7 @@ pub fn migrate_from_0_2_0_to_0_3_0(deps: DepsMut) -> StdResult<()> {
         lock_duration: old_config.lock_duration,
         reward_tokens: old_config.reward_tokens,
         deposits_enabled: old_config.deposits_enabled,
-        router: old_config.router,
+        router: cw_dex_router::helpers::CwDexRouter::new(&old_config.router.0),
         reward_liquidation_target: old_config.reward_liquidation_target,
         liquidity_helper: old_config.liquidity_helper,
         performance_fee: FeeConfig {
@@ -25,6 +26,21 @@ pub fn migrate_from_0_2_0_to_0_3_0(deps: DepsMut) -> StdResult<()> {
 
     // Store the new config
     CONFIG.save(deps.storage, &config)?;
+
+    Ok(())
+}
+
+#[allow(deprecated)]
+pub fn migrate_from_0_3_0_to_0_4_0(deps: DepsMut, incentives_contract: Addr) -> StdResult<()> {
+    let old_staking = locked_astroport_vault_0_2_0::state::STAKING.load(deps.storage)?;
+
+    let staking = AstroportStaking {
+        #[allow(deprecated)]
+        lp_token_addr: old_staking.lp_token_addr,
+        incentives: incentives_contract,
+    };
+
+    state::STAKING.save(deps.storage, &staking)?;
 
     Ok(())
 }